Test DistributorImpl, clean up DistributorImpl, JavaDoc
As specified in the title, test cases for all methods in DistributorImpl were added.

DistributorImpl was cleaned up because the implementation used to have repeated code in its methods. For all but one method where it was not possible (in a nice way), the methods now use a three step structure:

1. Obtaining the Sessions
2. Filtering the Sessions
3. Sending the specified object to the filtered Sessions

JavaDoc was added to all methods in the Distributor interface. The general interface description was left empty since I was unsure what one should write.